+++ client/src/components/OAuth.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import GooglePathImage from "./Path Images/GooglePathImage.jsx";
-import { handleSignIn } from "../utils/dataFetcher.js";
+import { handleSignIn } from "../utils/authControl.js";
 
 function OAuthSignIn() {
   return (
